+++ anime_dl/sites/__init__.py
@@ -1,2 +1,3 @@
 import sites.crunchyroll
 import sites.funimation
+import sites.vrv